+++ ORA_DAL/Model/PositionsDM.cs
@@ -2,7 +2,7 @@
 {
     public class PositionsDM
     {
-        public int PositionId { get; set; }
+        public long PositionId { get; set; }
 
         public string PositionName { get; set; }
     }
